expenxes

+++ lib/model/expences.dart
@@ -1,6 +1,6 @@
 class Expences {
 final String title;
-final int money;
+final double money;
 final String date;
 
   Expences({required this.title, required this.money, required this.date});
